modularised the program
