+++ modules/banner.py
@@ -1,0 +1,14 @@
+
+
+banner = """
+    _     _____                           _             
+   (_)   |_   _|                         | |            
+    _ ___  | |  _ __  ___ _ __   ___  ___| |_ ___  _ __ 
+   | / __| | | | '_ \/ __| '_ \ / _ \/ __| __/ _ \| '__|
+   | \__ \_| |_| | | \__ \ |_) |  __/ (__| || (_) | |   
+   | |___/_____|_| |_|___/ .__/ \___|\___|\__\___/|_|   
+  _/ |                   | |                            
+ |__/                    |_|                            
+ 
+Developed by github.com/BDragisic and github.com/Gr4y-r0se
+          """
